set tooltip style
